Finishing main components
